+++ packages/solid-form/src/createFormHook.tsx
@@ -1,7 +1,11 @@
-import { createContext, splitProps, useContext } from 'solid-js'
-import { createForm } from './createForm'
+import {
+  createComponent,
+  createContext,
+  splitProps,
+  useContext,
+} from 'solid-js'
 import { createFieldGroup } from './createFieldGroup'
-import type { AppFieldExtendedSolidFieldGroupApi } from './createFieldGroup'
+import { createForm } from './createForm'
 import type {
   AnyFieldApi,
   AnyFormApi,
@@ -21,6 +25,7 @@ import type {
   ParentProps,
 } from 'solid-js'
 import type { FieldComponent } from './createField'
+import type { AppFieldExtendedSolidFieldGroupApi } from './createFieldGroup'
 import type { SolidFormExtendedApi } from './createForm'
 
 /**
@@ -357,7 +362,13 @@ export function createFormHook<
         <form.Field {...fieldProps}>
           {(field) => (
             <opts.fieldContext.Provider value={field}>
-              {childProps.children(Object.assign(field, opts.fieldComponents))}
+              {createComponent(
+                () =>
+                  childProps.children(
+                    Object.assign(field, opts.fieldComponents),
+                  ),
+                {},
+              )}
             </opts.fieldContext.Provider>
           )}
         </form.Field>